% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annPeak.R
\name{annPeak_bytss}
\alias{annPeak_bytss}
\title{annPeak_bytss}
\usage{
annPeak_bytss(peak.gr, tx, tssup = 1000, tssdown = 1000)
}
\arguments{
\item{peak.gr}{A peak GRanges. This can be get from Bed file of peak using bed2gr function in this package.}

\item{tx}{A data frame of transcript/gene annotation, contains columns 'txid','seqnames', 'start', 'end', 'strand'.}

\item{tssup}{Window size around TSS that will be defined as overlap_tss, and 0 distance will be given. Default is 1000, means must overlap within 1 kb upstream of TSS.}

\item{tssdown}{Window size around TSS that will be defined as overlap_tss, and 0 distance will be given. Default is 1000, means must overlap within 1 kb downstream of TSS.}
}
\value{
A dataframe of annotated peaks with peak and it's nearest TSS information.
}
\description{
Annotate peak by it's nearest transcription start site.
If peak overlaps multiple tss, random one will be selected.
}

