% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipMap.R
\name{bin}
\alias{bin}
\title{bin}
\usage{
bin(v, bin)
}
\arguments{
\item{v}{numeric vector}

\item{bin}{size of bin}
}
\value{
A vector
}
\description{
Bin the vector by select the first element of the vector
}

