% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipMap.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(v, bin)
}
\arguments{
\item{v}{numeric vector}

\item{bin}{size of bin}
}
\value{
A vector
}
\description{
Bin the vector by return the mean of the vector
}

