% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMX.R
\name{metaPlot}
\alias{metaPlot}
\title{metaPlot}
\usage{
metaPlot(mxls, titles, out.pdf = F, color = NA, center = "peak",
  label_win = "5 kb", ylim = NA)
}
\arguments{
\item{mxls}{A list of matrix generated by tssMap or peakMap,must be the same dimensions.}

\item{titles}{Figure title and output file name}

\item{out.pdf}{Save image as PDF or PNG. Default is PNG.}

\item{color}{line colors}

\item{center}{center type of the matrix, 'peak' or 'tss'. Default is 'peak'.}

\item{label_win}{label of the window size. Default is '5 kb'.}

\item{ylim}{y limit of the plot. Default is 0 to the maximum value of the matrix.}
}
\value{
plot
}
\description{
This function plots the mean of each column to give a summary profile of the matrix, which is usually a hat shape.
}

