% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annPeak.R
\name{nearest_peak}
\alias{nearest_peak}
\title{nearest_peak}
\usage{
nearest_peak(peak.gr, tx, tssup = 1000, tssdown = 1000)
}
\arguments{
\item{peak.gr}{A peak GRanges. This can be get from Bed file of peak using bed2gr function in this package.}

\item{tx}{A data frame of transcript/gene annotation.}

\item{tssup}{Window size around TSS that will be defined as overlap_tss, and 0 distance will be given. Default is 1000, means must overlap within 1 kb upstream of TSS.}

\item{tssdown}{Window size around TSS that will be defined as overlap_tss, and 0 distance will be given. Default is 1000, means must overlap within 1 kb downstream of TSS.}
}
\value{
A dataframe with gene and it's nearest peak information.
}
\description{
Find nearest peak of given genes's transcription start site.
If tss overlaps multiple peaks, random one will be selected.
}

