% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipMap.R
\name{peakMap}
\alias{peakMap}
\title{peakMap}
\usage{
peakMap(query, anndf, left = -5000, right = 5000, bin_width = 25,
  sortByWidth = FALSE)
}
\arguments{
\item{query}{A coverage RLE. This can be get from a bigWig file using bw2coverage function in this package. BigWig file can be generated from Bam file using bamtools and bedtools.
Or a peak GRanges. This can be get from Bed file of peak using bed2gr function in this package. If the query is a peak GRanges, the output will be a enrichment profile instead of a coverage heatmap.}

\item{anndf}{A data frame contains columns 'seqnames', 'peak_start', 'peak_end', 'strand'. The function will find peak center. 'strand' can be '*'. Flip strand if strand is "-".}

\item{left}{Specify the window around TSS to be calculate. Default is -5 kb to 5 kb.}

\item{right}{Specify the window around TSS to be calculate. Default is -5 kb to 5 kb.}

\item{bin_width}{Specify the bin size to smooth and compress the window. Default is 25.}

\item{sortByWidth}{If sort by the peak width or not. Default if FALSE.}
}
\value{
A matrix with each line as a region around each peak centre and each column is a bin.
}
\description{
Calculate chip seq matrix around the centre of given peaks.
}

