% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMX.R
\name{plotMX}
\alias{plotMX}
\title{plotMX}
\usage{
plotMX(mxls, output, cutoff = 1, color = c("white", "blue"))
}
\arguments{
\item{mxls}{A list of matrix generated by tssMap or peakMap,must be the same dimensions.}

\item{output}{A PDF file name for image to save.}

\item{cutoff}{Color saturate value. Default is 1. The color will change between 0 to 1 and saturate above 1.}

\item{color}{The color to be used in the image.}
}
\value{
PDF images of the matrix and color bar. For enrichment profile, color bar is useless.
}
\description{
This function plots the matrix calculated by tssMap or peakMap.
}

