% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annPeak.R
\name{plotPie}
\alias{plotPie}
\title{plotPie}
\usage{
plotPie(annp, titles, out.pdf = F, color = c("yellow2", "orange", "red",
  "red4"))
}
\arguments{
\item{annp}{A dataframe of annPeak output.}

\item{titles}{Figure title and output file name}

\item{out.pdf}{Save image as PDF or PNG. Default is PNG.}

\item{color}{color in image.}
}
\value{
A image of PNG or PDF.
}
\description{
plot pie chart of 'peak_to_feature' distribution of annPeak results.
}

