% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annPeak.R
\name{plotPie_tss}
\alias{plotPie_tss}
\title{plotPie_tss}
\usage{
plotPie_tss(annp_bytss, titles, out.pdf = F, color = rev(brewer.pal(5,
  "Blues")))
}
\arguments{
\item{annp_bytss}{A dataframe of annPeak_bytss output.}

\item{titles}{Figure title and output file name}

\item{out.pdf}{Save image as PDF or PNG. Default is PNG.}

\item{color}{color in image.}
}
\value{
A image of PNG or PDF.
}
\description{
plot pie chart of 'distance_center_to_tss' distribution of annPeak_bytss results.
}

